+++ docs/gallery/general/extensions.py
@@ -248,16 +248,11 @@ io.close()
 # .. note::
 #
 #     For more information on writing NWB files, see :ref:`basic_writing`.
-
-####################
-# By default, PyNWB does not use the namespaces cached in a file--you must
-# explicitly specify this. This behavior is enabled by the *load_namespaces*
-# argument to the :py:class:`~pynwb.NWBHDF5IO` constructor.
-
-with NWBHDF5IO("cache_spec_example.nwb", mode="r", load_namespaces=True) as io:
-    nwbfile = io.read()
-
-####################
+#
+# By default, if a namespace is not already loaded, PyNWB loads the namespace cached in
+# the file. To disable this, set ``load_namespaces=False`` in the
+# :py:class:`~pynwb.NWBHDF5IO` constructor.
+#
 # .. _MultiContainerInterface:
 #
 # Creating and using a custom MultiContainerInterface